feat: move content to another page
